--- a/src/JustEat.StatsD/IStatsDTransport.cs
+++ b/src/JustEat.StatsD/IStatsDTransport.cs
@@ -1,7 +1,9 @@
+using System;
+
 namespace JustEat.StatsD
 {
     public interface IStatsDTransport
     {
-        void Send(string metric);
+        void Send(ReadOnlySpan<byte> metric);
     }
 }