--- a/src/JustEat.StatsD/IpTransport.cs
+++ b/src/JustEat.StatsD/IpTransport.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Buffers;
 using System.Net.Sockets;
 using System.Text;
 using JustEat.StatsD.EndpointLookups;
@@ -17,12 +18,15 @@ namespace JustEat.StatsD
         public void Send(string metric)
         {
             var endpoint = _endpointSource.GetEndpoint();
-            var bytes = Encoding.UTF8.GetBytes(metric);
+            var rent = ArrayPool<byte>.Shared.Rent(Encoding.UTF8.GetMaxByteCount(metric.Length));
+            var bytes = Encoding.UTF8.GetBytes(metric, rent);
 
             using (var socket = CreateSocket())
             {
-                socket.SendTo(bytes, endpoint);
+                socket.SendTo(rent, bytes, SocketFlags.None, endpoint);
             }
+
+            ArrayPool<byte>.Shared.Return(rent);
         }
 
         private static Socket CreateSocket()