This is a draft of string formatting with spans. Benchamrk.Net shows 0b allocated for formatting and constant 800b for transporting stats. 

Is there an option for such kind of change to get merged?