@slang25 2018-07-08
### FormatterBenchmark
        Method |      Mean |      Error |    StdDev |  Gen 0 | Allocated |
-------------- |----------:|-----------:|----------:|-------:|----------:|
  IncrementBy1 |  20.75 ns |   3.471 ns | 0.1961 ns | 0.0127 |      80 B |
 IncrementBy12 |  79.92 ns |  25.707 ns | 1.4525 ns | 0.0279 |     176 B |
     Decrement |  89.07 ns |  59.937 ns | 3.3866 ns | 0.0279 |     176 B |
         Event |  21.25 ns |   8.006 ns | 0.4523 ns | 0.0127 |      80 B |
         Gauge | 186.38 ns | 161.779 ns | 9.1408 ns | 0.0165 |     104 B |
        Timing |  90.21 ns |  19.432 ns | 1.0979 ns | 0.0304 |     192 B |

### StatSendingBenchmark
 Method |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
------- |----------:|----------:|----------:|-------:|----------:|
 RunUdp | 103.32 us | 50.502 us | 2.8535 us | 0.1221 |     800 B |
  RunIp |  74.36 us |  6.203 us | 0.3505 us | 0.1221 |     800 B |