@AnthonySteele 2018-07-09
What are the before and after metrics from benchmarkDotNet, on the same machine?

What are the platforms supported by this code? We currently target `net451` and `netstandard2.0`.  Though we could increment or drop the older platform, but only if we do a new major version.