@dv00d00 2018-07-09
I kept the old formatter and benchamrks for it. Span based results are:

|Method |      Mean |     Error |     StdDev |    Median | Allocated |
|-------------- |----------:|----------:|-----------:|----------:|----------:|
|  IncrementBy1 |  99.53 ns | 4.2938 ns | 12.2505 ns |  93.66 ns |       0 B |
| IncrementBy12 | 135.96 ns | 2.7377 ns |  5.4674 ns | 135.36 ns |       0 B |
|     Decrement | 146.06 ns | 1.8600 ns |  1.6488 ns | 146.12 ns |       0 B |
|         Event |  88.29 ns | 1.1799 ns |  1.1037 ns |  88.12 ns |       0 B |
|         Gauge | 127.66 ns | 0.6875 ns |  0.5368 ns | 127.78 ns |       0 B |
|        Timing | 133.87 ns | 2.6460 ns |  2.9410 ns | 132.76 ns |       0 B |