@dv00d00 2018-07-09
String-based formatter:

|Method |      Mean |     Error |   StdDev |    Median |  Gen 0 | Allocated |
|-------------- |----------:|----------:|---------:|----------:|-------:|----------:|
|  IncrementBy1 |  29.58 ns | 0.6804 ns | 1.118 ns |  29.26 ns | 0.0508 |      80 B |
| IncrementBy12 | 110.60 ns | 1.5780 ns | 1.318 ns | 110.43 ns | 0.1118 |     176 B |
|     Decrement | 120.23 ns | 2.0964 ns | 1.751 ns | 120.35 ns | 0.1118 |     176 B |
|         Event |  31.12 ns | 0.7074 ns | 1.775 ns |  30.39 ns | 0.0508 |      80 B |
|         Gauge | 247.25 ns | 4.9498 ns | 7.559 ns | 244.57 ns | 0.0658 |     104 B |
|        Timing | 126.62 ns | 2.0130 ns | 1.681 ns | 126.21 ns | 0.1218 |     192 B |