@dv00d00 2018-07-09
@AnthonySteele to get full benefit out of it the proper choice would be netcoreapp2.1+, and probably .net 4.8 if it gets native spans.

It still might be worth doing on a slowspan implementations (.net and netcoreapp < 2.1). I am not entirely sure if span based apis will work there, I will investigate.