@AnthonySteele 2018-07-09
My main concern is that the span run seems to be slower. [And if the benchmark is using 2.1, it should use the fast spans?](https://github.com/justeat/JustEat.StatsD/pull/77/files#diff-64354ac04d2b52d7a64e6a8950aebe38R5)