@dv00d00 2018-07-09
It looks like it is slower due to UTF8.GetBytes is called at formatter level, previously UTF8 get bytes was called before sending the packet over the wire