@dv00d00 2018-07-09
Previous implementation benchmark on my machine: 

|Method |     Mean |    Error |   StdDev |   Median | Allocated |
|------- |---------:|---------:|---------:|---------:|----------:|
| RunUdp | 832.3 us | 19.88 us | 56.07 us | 815.3 us |   1.16 KB |
|  RunIp | 774.6 us | 15.19 us | 21.30 us | 766.6 us |   1.16 KB |

This:

|Method |     Mean |     Error |    StdDev | Allocated |
|------- |---------:|----------:|----------:|----------:|
| RunUdp | 783.8 us | 12.176 us | 10.167 us |     800 B |
|  RunIp | 766.3 us |  8.581 us |  6.700 us |     800 B |