@dv00d00 2018-07-09
I will run benchmarks from my laptop, for some reason, it performs 8 times slower than a desktop.
After the latest change to transport with Connect-Send methods performance degraded approximately by 5%

It feels like major perf benefit would come from Socket's pooling, via something similar to HttpClientFactory but for sockets.

Perf boost from utf8 parser writing into stackalocked span is only slightly faster than String.Format, there might be some perf boost potential hidden there as well.

Reducing stacklocked buffer size to something dependant on payload helped a bit, will scratch it there as well.