@martincostello 2018-07-10
A quick two-pence thought into this PR (though I don't have time to properly review it until tomorrow), I'm concerned that in the rush to jump on the new-shiny-thing of Spans, we'd seem to be forcing the library to only support the latest version of .NET Core/Standard, which would suddenly leave users with slower release trains a bit left out in the dark for bugs fixes etc.