@slang25 2018-07-10
It could be `#if`ed to keep support for the current TFMs, or the other existing TFMs could take a dependency on `System.Memory`, which polyfill with "slow Span", which will still be crazy quick with this.

I vote for minimum impact to existing TFMs, so `#if` where needed.