@dv00d00 2018-07-10
Several updates:

I've added a benchmark with empty transport implementations. In them, span based version runs 33 % faster. But the absolute time change is in a range of milliseconds. The really slow part is a socket creation and IO it produces.

Average time to create a socket and send a packet via it is approximately 100ms on my machine. Perf boost from using non-allocating formatting falls into a range of standard deviation of sending the actual data over the wire.

I will focus on separating new formatter from existing stuff to make it optional and keep compatibility with older versions of .net.