@dv00d00 2018-07-10
So I've tried to add socket pooling for UDP transport

|           Method | Categories |      Mean |     Error |    StdDev | Scaled |  Gen 0 | Allocated |
|----------------- |----------- |----------:|----------:|----------:|-------:|-------:|----------:|
|           RunUdp |        UDP | 315.91 us | 49.740 us | 2.8104 us |   1.00 | 0.9766 |    4029 B |
| RunUdpSpanPooled |        UDP |  54.18 us |  6.273 us | 0.3544 us |   0.17 |      - |       0 B |
|                  |            |           |           |           |        |        |           |
|            RunIp |        TCP | 290.65 us | 22.373 us | 1.2641 us |   1.00 | 0.9766 |    4045 B |
|        RunIpSpan |        TCP | 284.13 us | 26.288 us | 1.4853 us |   0.98 | 0.4883 |    2208 B |