@martincostello 2018-07-10
I’ll look at the results again tomorrow, but socket-overhead is something that’s come up internally. We have a couple of apps using their own socket-pooling implementation, but the CPU overhead wasn’t quite quantifiable enough in benchmarks to make the case for it being the default implementation here in the main library.