@AnthonySteele 2018-07-11
I agree with Martin: the Spans have a lot of complexity, only run on a few of the frameworks that we use, produce small gains on something that's not the bottleneck.