@AnthonySteele 2018-07-11
We had pooled resources, but it wasn't helping performance. See https://github.com/justeat/JustEat.StatsD/pull/75/files
Possibly we weren't doing it well or in the right place and it can be re-introduced with more modern platform support?