@slang25 2018-07-11
@martincostello https://youtu.be/LCV-JNOkZC4?t=6m5s
This clip pretty much covers my understanding. It's worth watching 10 minutes from there, my comment gets read out and your tweet and gets shown too.