@martincostello 2018-07-11
I'd be much happier with a separate package until it's more reasonable to bring up the drawbridge on older TFMs, but even if we add an extra TFM and use `#if` now, we're still creating divergence as you need to make a breaking change to the core interface (for the 3rd time this year?) so you can pass in a span.