@dv00d00 2018-07-17
Closing it cause it definitely would not be merged, will still refer to in future PRs.